Refactor and prune generated code
